Cleans up the plugin implementation to confirm it runs in production
We tried skipping the internal autoloader, in case the consuming site had already loaded
the plugin classes - but that seems not to work.

This also implements a build branch, populated by a GitHub Action, in order to populate
our internal dependencies in the vendor/ directory (which are usually .gitignored for
local use, but then are needed by consuming sites it seems).

Along the way we beef up the linting on GitHub Actions as well.